# Conflicts:
#	src/main/antlr/in.handyman.raven/gmr/Raven.g4
#	src/main/java/in/handyman/raven/lib/model/agentic/paper/filter/AgenticPaperFilterConsumerProcess.java

+++ src/main/java/in/handyman/raven/lib/model/documentEyeCue/DocumentEyeCueResponse.java
@@ -22,6 +22,6 @@ public class DocumentEyeCueResponse {
     private String processedPdfPath;
     private String processedPdfBase64;
     private String processedPdfChecksum;
-    private int docEyeCueDurationMs;
+    private Long docEyeCueDurationMs;
     private String errorMessage;
 }
